requirements.txt
pytest
gradio
black
numpy
tensorflow
pytesseract
opencv-python
python-resize-image